# docs/api-reference/app-client-openapi.json
{
  "components": {
    "responses": {
      "Error401": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Missing or invalid authentication credentials."
      },
...
    },
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SetUserLanguageRequest": {
        "properties": {
          "language": {
            "title": "Language",
            "type": "string"
          }
        },
        "required": [
          "language"
        ],
        "title": "SetUserLanguageRequest",
        "type": "object"
      },
...
      "UserDataExportResponse": {
        "properties": {
          "account": {
            "additionalProperties": true,
            "title": "Account",
            "type": "object"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          },
          "subscription": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription"
          },
          "usage": {
            "additionalProperties": true,
            "title": "Usage",
            "type": "object"
          }
        },
        "required": [
          "schema_version"
        ],
        "title": "UserDataExportResponse",
        "type": "object"
      },
      "UserLanguageResponse": {
        "properties": {
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "title": "UserLanguageResponse",
        "type": "object"
      },
...
      "UserStatusResponse": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "UserStatusResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "firebaseBearer": {
        "bearerFormat": "Firebase ID token",
        "description": "Send `Authorization: Bearer <firebase_id_token>`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Intentive"
    },
    "description": "First-party Intentive app access to transient conversation and Memory compute and product services.",
    "license": {
      "name": "MIT",
      "url": "https://github.com/sruj75/knowledge-athlete/blob/main/LICENSE"
    },
    "title": "Intentive App Client API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/users/delete-account": {
      "delete": {
        "operationId": "delete_account_v1_users_delete_account_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Platform",
            "required": false,
            "schema": {
              "title": "X-App-Platform",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Device-Id-Hash",
            "required": false,
            "schema": {
              "title": "X-Device-Id-Hash",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Version",
            "required": false,
            "schema": {
              "title": "X-App-Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "summary": "Delete Account",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/users/export": {
      "get": {
        "description": "Export retained server-owned account and entitlement metadata.",
        "operationId": "export_all_user_data_v1_users_export_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Platform",
            "required": false,
            "schema": {
              "title": "X-App-Platform",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Device-Id-Hash",
            "required": false,
            "schema": {
              "title": "X-Device-Id-Hash",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Version",
            "required": false,
            "schema": {
              "title": "X-App-Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDataExportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "summary": "Export All User Data",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/users/language": {
      "get": {
        "description": "Get the user's preferred language.",
        "operationId": "get_user_language_v1_users_language_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Platform",
            "required": false,
            "schema": {
              "title": "X-App-Platform",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Device-Id-Hash",
            "required": false,
            "schema": {
              "title": "X-Device-Id-Hash",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Version",
            "required": false,
            "schema": {
              "title": "X-App-Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLanguageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "summary": "Get User Language",
        "tags": [
          "v1"
        ]
      },
      "patch": {
        "description": "Set the user's preferred language (e.g., 'en', 'vi', etc.).",
        "operationId": "set_user_language_v1_users_language_patch",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Platform",
            "required": false,
            "schema": {
              "title": "X-App-Platform",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Device-Id-Hash",
            "required": false,
            "schema": {
              "title": "X-Device-Id-Hash",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Version",
            "required": false,
            "schema": {
              "title": "X-App-Version",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUserLanguageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserLanguageUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "summary": "Set User Language",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/users/me/llm-usage/total": {
      "get": {
        "operationId": "get_total_llm_cost_v1_users_me_llm_usage_total_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Platform",
            "required": false,
            "schema": {
              "title": "X-App-Platform",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Device-Id-Hash",
            "required": false,
            "schema": {
              "title": "X-Device-Id-Hash",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Version",
            "required": false,
            "schema": {
              "title": "X-App-Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LlmTotalCostResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "summary": "Get Total Llm Cost",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/me/paywall": {
      "get": {
        "description": "Trial-paywall status for the calling user on the given platform.\n\nUsed by canonical desktop provider paths to decide whether to admit\npaid LLM / TTS traffic. Mirrors the exact semantics of\n`is_trial_paywalled`: basic plan + Firebase Auth account >3d old + platform\nin {macos, desktop}. Mobile platforms always\nreturn `paywalled=false`.\n\nPlatform comes from `X-App-Platform` header (preferred) or `platform`\nquery param (fallback). Unknown / missing platforms are never paywalled.",
        "operationId": "get_user_paywall_status_v1_users_me_paywall_get",
        "parameters": [
          {
            "in": "query",
            "name": "platform",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          },
          {
            "in": "header",
            "name": "X-App-Platform",
            "required": false,
            "schema": {
              "title": "X-App-Platform",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Device-Id-Hash",
            "required": false,
            "schema": {
              "title": "X-Device-Id-Hash",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Version",
            "required": false,
            "schema": {
              "title": "X-App-Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaywallStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "summary": "Get User Paywall Status",
        "tags": [
          "users"
        ]
      }
    },
    "/v1/users/me/subscription": {
      "get": {
        "description": "Return the server-owned subscription, usage, and billing capability.",
        "operationId": "get_user_subscription_endpoint_v1_users_me_subscription_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-App-Platform",
            "required": false,
            "schema": {
              "title": "X-App-Platform",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-App-Version",
            "required": false,
            "schema": {
              "title": "X-App-Version",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Device-Id-Hash",
            "required": false,
            "schema": {
              "title": "X-Device-Id-Hash",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSubscriptionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "summary": "Get User Subscription Endpoint",
        "tags": [
          "v1"
        ]
      }
    },
...
  },
...
}